Update shadow system (#2059)
* Fix shadow bias when the atkas uses a high rescale level

* Cleanup shadow files

* More cleanup

* Use cascade settings from volumes instead of light datas

* Fixed a shadow atlas overflow when using global atlas scaling
