+++ com.unity.render-pipelines.high-definition/Editor/Lighting/AdditionalShadowDataEditor.cs
@@ -5,7 +5,7 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
 {
     [CanEditMultipleObjects]
     // Disable HDRP custom editor to display full shadow settings (only for dev purpose, reset for pr)
-    // [CustomEditor(typeof(AdditionalShadowData))]
+    [CustomEditor(typeof(AdditionalShadowData))]
     class AdditionalShadowDataEditor : Editor
     {
         public override void OnInspectorGUI()
