--- com.unity.render-pipelines.high-definition/Editor/RenderPipeline/Settings/SerializedShadowInitParameters.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a0e52dc13dec5924a9ac8a3f279cea36
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
