--- com.unity.render-pipelines.high-definition/Editor/RenderPipeline/Settings/ShadowInitParametersUI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5182f05875d6503459eda16786882b44
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
