+++ com.unity.render-pipelines.high-definition/Runtime/Lighting/LightLoop/Deferred.compute
@@ -65,8 +65,6 @@
 #   pragma warning( disable : 4714 ) // sum of temp registers and indexable temp registers times 256 threads exceeds the recommended total 16384.  Performance may be reduced at kernel
 #endif
 
-// #define USE_CORE_SHADOW_SYSTEM
-
 #define UNITY_SINGLE_PASS_STEREO 1
 
 #define LIGHTLOOP_TILE_PASS 1
