+++ com.unity.render-pipelines.high-definition/Runtime/Lighting/LightLoop/LightLoop.cs
@@ -28,161 +28,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
             return value.localToWorld.GetColumn(0);
         }
     }
-
-    class ShadowSetup : IDisposable
-    {
-        // shadow related stuff
-        const int k_MaxShadowDataSlots              = 64;
-        const int k_MaxPayloadSlotsPerShadowData    =  4;
-        ShadowmapBase[]         m_Shadowmaps;
-        ShadowManager           m_ShadowMgr;
-        static ComputeBuffer    s_ShadowDataBuffer;
-        static ComputeBuffer    s_ShadowPayloadBuffer;
-
-        public static GPUShadowType HDShadowLightType(Light l)
-        {
-            // We only process light with additional data
-            var ald = l.GetComponent<HDAdditionalLightData>();
-
-            if (ald == null)
-            {
-                return ShadowRegistry.ShadowLightType(l);
-            }
-
-            GPUShadowType shadowType = GPUShadowType.Unknown;
-
-            switch (ald.lightTypeExtent)
-            {
-                case LightTypeExtent.Punctual:
-                    shadowType = ShadowRegistry.ShadowLightType(l);
-                    break;
-
-                    // Area and projector not supported yet
-            }
-
-            return shadowType;
-        }
-
-        public ShadowSetup(RenderPipelineResources resources, ShadowInitParameters shadowInit, ShadowSettings shadowSettings, out IShadowManager shadowManager)
-        {
-            s_ShadowDataBuffer      = new ComputeBuffer(k_MaxShadowDataSlots, System.Runtime.InteropServices.Marshal.SizeOf(typeof(ShadowData)));
-            s_ShadowPayloadBuffer   = new ComputeBuffer(k_MaxShadowDataSlots * k_MaxPayloadSlotsPerShadowData, System.Runtime.InteropServices.Marshal.SizeOf(typeof(ShadowPayload)));
-            ShadowAtlas.AtlasInit atlasInit;
-            atlasInit.baseInit.width           = (uint)shadowInit.shadowAtlasWidth;
-            atlasInit.baseInit.height          = (uint)shadowInit.shadowAtlasHeight;
-            atlasInit.baseInit.slices          = 1;
-            atlasInit.baseInit.shadowmapBits   = shadowInit.shadowMap16Bit ? 16u : 32u;
-            atlasInit.baseInit.shadowmapFormat = RenderTextureFormat.Shadowmap;
-            atlasInit.baseInit.samplerState    = SamplerState.Default();
-            atlasInit.baseInit.comparisonSamplerState = ComparisonSamplerState.Default();
-            atlasInit.baseInit.clearColor      = new Vector4(0.0f, 0.0f, 0.0f, 0.0f);
-            atlasInit.baseInit.maxPayloadCount = 0;
-            atlasInit.baseInit.shadowSupport   = ShadowmapBase.ShadowSupport.Directional | ShadowmapBase.ShadowSupport.Point | ShadowmapBase.ShadowSupport.Spot;
-            atlasInit.shaderKeyword            = null;
-            atlasInit.shadowClearShader         = resources.shaders.shadowClearPS;
-            atlasInit.shadowBlurMoments         = resources.shaders.shadowBlurMomentsCS;
-
-            /*
-            // Code kept here for reference if we want to add VSM/MSM later on
-            var varianceInit = atlasInit;
-            varianceInit.baseInit.shadowmapFormat = ShadowVariance.GetFormat(false, false, true);
-            var varianceInit2 = varianceInit;
-            varianceInit2.baseInit.shadowmapFormat = ShadowVariance.GetFormat(true, true, false);
-            var varianceInit3 = varianceInit;
-            varianceInit3.baseInit.shadowmapFormat = ShadowVariance.GetFormat(true, false, true);
-            m_Shadowmaps = new ShadowmapBase[] { new ShadowAtlas(ref atlasInit), new ShadowVariance(ref varianceInit), new ShadowVariance(ref varianceInit2), new ShadowVariance(ref varianceInit3) };
-            */
-
-            m_Shadowmaps = new ShadowmapBase[] { new ShadowAtlas(ref atlasInit) };
-
-            ShadowContext.SyncDel syncer = (ShadowContext sc) =>
-                {
-                    // update buffers
-                    uint offset, count;
-                    ShadowData[] sds;
-                    sc.GetShadowDatas(out sds, out offset, out count);
-                    Debug.Assert(offset == 0);
-
-                    // Avoid overflowing the compute buffer
-                    if (sds.Length <= k_MaxShadowDataSlots)
-                        s_ShadowDataBuffer.SetData(sds);   // unfortunately we can't pass an offset or count to this function
-                    
-                    ShadowPayload[] payloads;
-                    sc.GetPayloads(out payloads, out offset, out count);
-                    Debug.Assert(offset == 0);
-                    
-                    // Avoid overflowing the compute buffer
-                    if (payloads.Length <= k_MaxShadowDataSlots)
-                        s_ShadowPayloadBuffer.SetData(payloads);
-                };
-
-
-            // binding code. This needs to be in sync with ShadowContext.hlsl
-            ShadowContext.BindDel binder = (ShadowContext sc, CommandBuffer cb, ComputeShader computeShader, int computeKernel) =>
-                {
-                    uint offset, count;
-                    RenderTargetIdentifier[] tex;
-                    sc.GetTex2DArrays(out tex, out offset, out count);
-
-                    // bind buffers
-                    cb.SetGlobalBuffer(HDShaderIDs._ShadowDatasExp, s_ShadowDataBuffer);
-                    cb.SetGlobalBuffer(HDShaderIDs._ShadowPayloads, s_ShadowPayloadBuffer);
-                    // bind textures
-                    cb.SetGlobalTexture(HDShaderIDs._ShadowmapExp_PCF, tex[0]);
-                    // Code kept here for reference if we want to add VSM/MSM later on
-                    //cb.SetGlobalTexture(HDShaderIDs._ShadowmapExp_VSM_0, tex[1]);
-                    //cb.SetGlobalTexture(HDShaderIDs._ShadowmapExp_VSM_1, tex[2]);
-                    //cb.SetGlobalTexture(HDShaderIDs._ShadowmapExp_VSM_2, tex[3])
-
-                    // TODO: Currently samplers are hard coded in ShadowContext.hlsl, so we can't really set them here
-                };
-
-            ShadowContext.CtxtInit scInit;
-            scInit.storage.maxShadowDataSlots        = k_MaxShadowDataSlots;
-            scInit.storage.maxPayloadSlots           = k_MaxShadowDataSlots * k_MaxPayloadSlotsPerShadowData;
-            scInit.storage.maxTex2DArraySlots        = 4;
-            scInit.storage.maxTexCubeArraySlots      = 0;
-            scInit.storage.maxComparisonSamplerSlots = 1;
-            scInit.storage.maxSamplerSlots           = 4;
-            scInit.dataSyncer                        = syncer;
-            scInit.resourceBinder                    = binder;
-
-            ShadowManager.ShadowBudgets budgets;
-            budgets.maxPointLights       = shadowInit.maxPointLightShadows;
-            budgets.maxSpotLights        = shadowInit.maxSpotLightShadows;
-            budgets.maxDirectionalLights = shadowInit.maxDirectionalLightShadows;
-
-            m_ShadowMgr = new ShadowManager(shadowSettings, ref scInit, ref budgets, m_Shadowmaps);
-            // set global overrides - these need to match the override specified in LightLoop/Shadow.hlsl
-            bool useGlobalOverrides = true;
-            m_ShadowMgr.SetGlobalShadowOverride(GPUShadowType.Point        , ShadowAlgorithm.PCF, ShadowVariant.V3, shadowInit.shadowMap16Bit ? ShadowPrecision.Low : ShadowPrecision.High, useGlobalOverrides);
-            m_ShadowMgr.SetGlobalShadowOverride(GPUShadowType.Spot         , ShadowAlgorithm.PCF, ShadowVariant.V3, shadowInit.shadowMap16Bit ? ShadowPrecision.Low : ShadowPrecision.High, useGlobalOverrides);
-            m_ShadowMgr.SetGlobalShadowOverride(GPUShadowType.Directional  , ShadowAlgorithm.PCF, ShadowVariant.V3, shadowInit.shadowMap16Bit ? ShadowPrecision.Low : ShadowPrecision.High, useGlobalOverrides);
-
-            m_ShadowMgr.SetShadowLightTypeDelegate(HDShadowLightType);
-
-            shadowManager = m_ShadowMgr;
-        }
-
-        public void Dispose()
-        {
-            if (m_Shadowmaps != null)
-            {
-                foreach (var shadowMap in m_Shadowmaps)
-                {
-                    (shadowMap as ShadowAtlas).Dispose();
-                }
-                m_Shadowmaps = null;
-            }
-            m_ShadowMgr = null;
-
-            if (s_ShadowDataBuffer != null)
-                s_ShadowDataBuffer.Release();
-            if (s_ShadowPayloadBuffer != null)
-                s_ShadowPayloadBuffer.Release();
-        }
-    }
-
+    
     //-----------------------------------------------------------------------------
     // structure definition
     //-----------------------------------------------------------------------------
@@ -446,8 +292,6 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
         static ComputeBuffer s_PerTileLogBaseTweak = null;
         static ComputeBuffer s_GlobalLightListAtomic = null;
 
-        static bool useNewShadowSystem = true;
-
         public enum ClusterPrepassSource : int
         {
             None = 0,
@@ -488,7 +332,6 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
         // Following is an array of material of size eight for all combination of keyword: OUTPUT_SPLIT_LIGHTING - LIGHTLOOP_TILE_PASS - SHADOWS_SHADOWMASK - USE_FPTL_LIGHTLIST/USE_CLUSTERED_LIGHTLIST - DEBUG_DISPLAY
         Material[] m_deferredLightingMaterial;
         Material m_DebugViewTilesMaterial;
-        Material m_DebugShadowMapMaterial;
         Material m_DebugHDShadowMapMaterial;
         Material m_DebugLightVolumeMaterial;
         Material m_CubeToPanoMaterial;
@@ -505,38 +348,23 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
         public Light GetCurrentSunLight() { return m_CurrentSunLight; }
 
         // shadow related stuff
-        FrameId                             m_FrameId = new FrameId();
-        ShadowSetup                         m_ShadowSetup; // doesn't actually have to reside here, it would be enough to pass the IShadowManager in from the outside
-        IShadowManager                      m_ShadowMgr;
-        HDShadowManager                     m_NewShadowManager;
-        List<int>                           m_ShadowRequests = new List<int>();
-        Dictionary<int, int>                m_ShadowIndices = new Dictionary<int, int>();
+        HDShadowManager                     m_ShadowManager;
+        HDShadowInitParameters              m_ShadowInitParameters;
 
-        /// HD Shadow stuff
-        HDShadowInitParameters              m_HDShadowInitParameters;
         // Used to shadow shadow maps with use selection enabled in the debug menu
         int m_DebugSelectedLightShadowIndex;
         int m_DebugSelectedLightShadowCount;
 
-        void InitShadowSystem(HDRenderPipelineAsset hdAsset, ShadowSettings shadowSettings)
+        void InitShadowSystem(HDRenderPipelineAsset hdAsset)
         {
-            var shadowInitParams = hdAsset.GetRenderPipelineSettings().shadowInitParams;
-            m_HDShadowInitParameters = hdAsset.GetRenderPipelineSettings().hdShadowInitParams;
-            m_ShadowSetup = new ShadowSetup(hdAsset.renderPipelineResources, shadowInitParams, shadowSettings, out m_ShadowMgr);
-            m_NewShadowManager = new HDShadowManager(m_HDShadowInitParameters.shadowAtlasWidth, m_HDShadowInitParameters.shadowAtlasHeight, m_HDShadowInitParameters.maxShadowRequests, m_HDShadowInitParameters.shadowMapsDepthBits, hdAsset.renderPipelineResources.shaders.shadowClearPS);
+            m_ShadowInitParameters = hdAsset.GetRenderPipelineSettings().hdShadowInitParams;
+            m_ShadowManager = new HDShadowManager(m_ShadowInitParameters.shadowAtlasWidth, m_ShadowInitParameters.shadowAtlasHeight, m_ShadowInitParameters.maxShadowRequests, m_ShadowInitParameters.shadowMapsDepthBits, hdAsset.renderPipelineResources.shaders.shadowClearPS);
         }
 
         void DeinitShadowSystem()
         {
-            if (m_ShadowSetup != null)
-            {
-                m_ShadowSetup.Dispose();
-                m_ShadowSetup = null;
-                m_ShadowMgr = null;
-            }
-
-            m_NewShadowManager.Dispose();
-            m_NewShadowManager = null;
+            m_ShadowManager.Dispose();
+            m_ShadowManager = null;
         }
 
         int GetNumTileFtplX(HDCamera hdCamera)
@@ -573,12 +401,11 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
             return (outputSplitLighting) | (lightLoopTilePass << 1) | (shadowMask << 2) | (debugDisplay << 3);
         }
 
-        public void Build(HDRenderPipelineAsset hdAsset, ShadowSettings shadowSettings, IBLFilterGGX iblFilterGGX)
+        public void Build(HDRenderPipelineAsset hdAsset, IBLFilterGGX iblFilterGGX)
         {
             m_Resources = hdAsset.renderPipelineResources;
 
             m_DebugViewTilesMaterial = CoreUtils.CreateEngineMaterial(m_Resources.shaders.debugViewTilesPS);
-            m_DebugShadowMapMaterial = CoreUtils.CreateEngineMaterial(m_Resources.shaders.debugShadowMapPS);
             m_DebugHDShadowMapMaterial = CoreUtils.CreateEngineMaterial(m_Resources.shaders.debugHDShadowMapPS);
             m_DebugLightVolumeMaterial = CoreUtils.CreateEngineMaterial(m_Resources.shaders.debugLightVolumePS);
             m_CubeToPanoMaterial = CoreUtils.CreateEngineMaterial(m_Resources.shaders.cubeToPanoPS);
@@ -702,7 +529,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
             Shader.EnableKeyword(punctualShadowKeywords[(int)shadowParams.punctualShadowQuality]);
             Shader.EnableKeyword(directionalShadowKeywords[(int)shadowParams.directionalShadowQuality]);
 
-            InitShadowSystem(hdAsset, shadowSettings);
+            InitShadowSystem(hdAsset);
         }
 
         public void Cleanup()
@@ -764,7 +591,6 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
             }
 
             CoreUtils.Destroy(m_DebugViewTilesMaterial);
-            CoreUtils.Destroy(m_DebugShadowMapMaterial);
             CoreUtils.Destroy(m_DebugHDShadowMapMaterial);
             CoreUtils.Destroy(m_CubeToPanoMaterial);
         }
@@ -933,7 +759,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
             return true;
         }
 
-        public bool GetDirectionalLightData(CommandBuffer cmd, ShadowSettings shadowSettings, GPULightType gpuLightType, VisibleLight light, Light lightComponent, HDAdditionalLightData additionalData, AdditionalShadowData additionalShadowData, int lightIndex, int shadowIndex, DebugDisplaySettings debugDisplaySettings, int sortedIndex)
+        public bool GetDirectionalLightData(CommandBuffer cmd, GPULightType gpuLightType, VisibleLight light, Light lightComponent, HDAdditionalLightData additionalData, AdditionalShadowData additionalShadowData, int lightIndex, int shadowIndex, DebugDisplaySettings debugDisplaySettings, int sortedIndex)
         {
             // Clamp light list to the maximum allowed lights on screen to avoid ComputeBuffer overflow
             if (m_lightList.directionalLights.Count >= k_MaxDirectionalLightsOnScreen)
@@ -975,24 +801,11 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
                 directionalLightData.cookieIndex = m_CookieTexArray.FetchSlice(cmd, lightComponent.cookie);
             }
             // fix up shadow information
-            if (useNewShadowSystem)
+            directionalLightData.shadowIndex = shadowIndex;
+            if (shadowIndex != -1)
             {
-                directionalLightData.shadowIndex = shadowIndex;
-                if (shadowIndex != -1)
-                {
-                    m_CurrentSunLight = lightComponent;
-                    m_CurrentShadowSortedSunLightIndex = sortedIndex;
-                }
-            }
-            else
-            {
-                int shadowIdx;
-                if (m_ShadowIndices.TryGetValue(lightIndex, out shadowIdx))
-                {
-                    directionalLightData.shadowIndex = shadowIdx;
-                    m_CurrentSunLight = lightComponent;
-                    m_CurrentShadowSortedSunLightIndex = sortedIndex;
-                }
+                m_CurrentSunLight = lightComponent;
+                m_CurrentShadowSortedSunLightIndex = sortedIndex;
             }
 
             directionalLightData.shadowMaskSelector = Vector4.zero;
@@ -1054,7 +867,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
             return 1.0f - Mathf.Clamp01(distanceToCamera * scale + bias);
         }
 
-        public bool GetLightData(CommandBuffer cmd, ShadowSettings shadowSettings, Camera camera, GPULightType gpuLightType,
+        public bool GetLightData(CommandBuffer cmd, HDShadowSettings shadowSettings, Camera camera, GPULightType gpuLightType,
             VisibleLight light, Light lightComponent, HDAdditionalLightData additionalLightData, AdditionalShadowData additionalshadowData,
             int lightIndex, int shadowIndex, ref Vector3 lightDimensions, DebugDisplaySettings debugDisplaySettings)
         {
@@ -1237,18 +1050,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
             }
 
             // fix up shadow information
-            if (useNewShadowSystem)
-            {
-                lightData.shadowIndex = shadowIndex;
-            }
-            else
-            {
-                int shadowIdx;
-                if (m_ShadowIndices.TryGetValue(lightIndex, out shadowIdx))
-                {
-                    lightData.shadowIndex = shadowIdx;
-                }
-            }
+            lightData.shadowIndex = shadowIndex;
 
             // Value of max smoothness is from artists point of view, need to convert from perceptual smoothness to roughness
             lightData.minRoughness = (1.0f - additionalLightData.maxSmoothness) * (1.0f - additionalLightData.maxSmoothness);
@@ -1664,28 +1466,13 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
 
         public int GetCurrentShadowCount()
         {
-            if (useNewShadowSystem)
-                return m_NewShadowManager.GetShadowRequestCount();
-            else
-                return m_ShadowRequests.Count;
-        }
-
-        public int GetShadowAtlasCount()
-        {
-            return (m_ShadowMgr == null) ? 0 : (int)m_ShadowMgr.GetShadowMapCount();
-        }
-
-        public int GetShadowSliceCount(uint atlasIndex)
-        {
-            return (m_ShadowMgr == null) ? 0 : (int)m_ShadowMgr.GetShadowMapSliceCount(atlasIndex);
+            return m_ShadowManager.GetShadowRequestCount();
         }
 
         public void UpdateCullingParameters(ref ScriptableCullingParameters cullingParams)
         {
-            if (useNewShadowSystem)
-                m_NewShadowManager.UpdateCullingParameters(ref cullingParams);
-            else
-            m_ShadowMgr.UpdateCullingParameters(ref cullingParams);
+            m_ShadowManager.UpdateCullingParameters(ref cullingParams);
+            
             // In HDRP we don't need per object light/probe info so we disable the native code that handles it.
             cullingParams.cullingFlags |= CullFlag.DisablePerObjectCulling;
         }
@@ -1727,7 +1514,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
         }
 
         // Return true if BakedShadowMask are enabled
-        public bool PrepareLightsForGPU(CommandBuffer cmd, HDCamera hdCamera, ShadowSettings shadowSettings, CullResults cullResults,
+        public bool PrepareLightsForGPU(CommandBuffer cmd, HDCamera hdCamera, CullResults cullResults,
             ReflectionProbeCullResults reflectionProbeCullResults, DensityVolumeList densityVolumes, DebugDisplaySettings debugDisplaySettings)
         {
             using (new ProfilingSample(cmd, "Prepare Lights For GPU"))
@@ -1747,6 +1534,8 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
 
                 var stereoEnabled = m_FrameSettings.enableStereo;
 
+                var hdShadowSettings = VolumeManager.instance.stack.GetComponent<HDShadowSettings>();
+
                 Vector3 camPosWS = camera.transform.position;
 
                 var worldToView = WorldToCamera(camera);
@@ -1756,76 +1545,13 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
                     worldToView = WorldToViewStereo(camera, Camera.StereoscopicEye.Left);
                     rightEyeWorldToView = WorldToViewStereo(camera, Camera.StereoscopicEye.Right);
                 }
-
-                // TODO: Only for dev purpose remove once the new shadow system is stable
-                if (useNewShadowSystem)
-                    Shader.DisableKeyword("USE_CORE_SHADOW_SYSTEM");
-                else
-                    Shader.EnableKeyword("USE_CORE_SHADOW_SYSTEM");
                 
                 // We must clear the shadow requests before checking if they are any visible light because we would have requests from the last frame executed in the case where we don't see any lights
-                if (useNewShadowSystem)
-                {
-                    m_NewShadowManager.Clear();
-                }
+                m_ShadowManager.Clear();
 
                 // Note: Light with null intensity/Color are culled by the C++, no need to test it here
                 if (cullResults.visibleLights.Count != 0 || cullResults.visibleReflectionProbes.Count != 0)
                 {
-                    // 0. deal with shadows
-                    {
-                        if (!useNewShadowSystem)
-                        {
-                            m_FrameId.frameCount++;
-                            // get the indices for all lights that want to have shadows
-                            m_ShadowRequests.Clear();
-                            m_ShadowRequests.Capacity = cullResults.visibleLights.Count;
-                            int lcnt = cullResults.visibleLights.Count;
-                            for (int i = 0; i < lcnt; ++i)
-                            {
-                                VisibleLight vl = cullResults.visibleLights[i];
-                                var lightComponent = vl.light;
-
-                                // This can happen for particle light which don't have a proper game object. We don't support shadows for them.
-                                if (lightComponent == null)
-                                    continue;
-
-                                if (lightComponent.shadows == LightShadows.None)
-                                    continue;
-
-                                AdditionalShadowData asd = lightComponent.GetComponent<AdditionalShadowData>();
-                                if (asd != null && asd.shadowDimmer > 0.0f)
-                                {
-                                    m_ShadowRequests.Add(i);
-
-                                    // Discover sun light and update cascade info from Volumes
-                                    // TODO: This should be moved to GetDirectionalLightData when we merge the two loops here.
-                                    // Careful it must still be done BEFORE the call to ProcessShadowRequests
-                                    if (vl.lightType == LightType.Directional && m_CurrentSunLight == null)
-                                    {
-                                        var hdShadowSettings = VolumeManager.instance.stack.GetComponent<HDShadowSettings>();
-                                        asd.SetShadowCascades(hdShadowSettings.cascadeShadowSplitCount, hdShadowSettings.cascadeShadowSplits, hdShadowSettings.cascadeShadowBorders);
-                                    }
-                                }
-                            }
-                            // pass this list to a routine that assigns shadows based on some heuristic
-                            uint shadowRequestCount = (uint)m_ShadowRequests.Count;
-                            //TODO: Do not call ToArray here to avoid GC, refactor API
-                            int[]   shadowRequests = m_ShadowRequests.ToArray();
-                            int[]   shadowDataIndices;
-
-                            m_ShadowMgr.ProcessShadowRequests(m_FrameId, cullResults, camera, ShaderConfig.s_CameraRelativeRendering != 0, cullResults.visibleLights,
-                                ref shadowRequestCount, shadowRequests, out shadowDataIndices);
-
-                            // update the visibleLights with the shadow information
-                            m_ShadowIndices.Clear();
-                            for (uint i = 0; i < shadowRequestCount; i++)
-                            {
-                                m_ShadowIndices.Add(shadowRequests[i], shadowDataIndices[i]);
-                            }
-                        }
-                    }
-
                     // 1. Count the number of lights and sort all lights by category, type and volume - This is required for the fptl/cluster shader code
                     // If we reach maximum of lights available on screen, then we discard the light.
                     // Lights are processed in order, so we don't discards light based on their importance but based on their ordering in visible lights list.
@@ -1923,16 +1649,8 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
                             }
                         }
 
-                        uint shadow;
-                        if (useNewShadowSystem)
-                        {
-                            // TODO: Sorting by shadows is certainly useless here so we don't do it
-                            shadow = 0;
-                        }
-                        else
-                            shadow = m_ShadowIndices.ContainsKey(lightIndex) ? 1u : 0;
                         // 5 bit (0x1F) light category, 5 bit (0x1F) GPULightType, 5 bit (0x1F) lightVolume, 1 bit for shadow casting, 16 bit index
-                        m_SortKeys[sortCount++] = (uint)lightCategory << 27 | (uint)gpuLightType << 22 | (uint)lightVolumeType << 17 | shadow << 16 | (uint)lightIndex;
+                        m_SortKeys[sortCount++] = (uint)lightCategory << 27 | (uint)gpuLightType << 22 | (uint)lightVolumeType << 17 | (uint)lightIndex;
                     }
 
                     CoreUnsafeUtils.QuickSort(m_SortKeys, 0, sortCount - 1); // Call our own quicksort instead of Array.Sort(sortKeys, 0, sortCount) so we don't allocate memory (note the SortCount-1 that is different from original call).
@@ -1969,15 +1687,15 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
 
                         int shadowIndex = -1;
                         // Manage shadow requests
-                        if (lightComponent.shadows != LightShadows.None && useNewShadowSystem)
+                        if (lightComponent.shadows != LightShadows.None)
                         {
                             Bounds bounds;
 
                             // Update light shadow requests only if the light affects at least one object
-                            if (shadowSettings.enabled && cullResults.GetShadowCasterBounds(lightIndex, out bounds))
+                            if (m_FrameSettings.enableShadow && cullResults.GetShadowCasterBounds(lightIndex, out bounds))
                             {
                                 int shadowRequestCount;
-                                shadowIndex = additionalLightData.UpdateShadowRequest(camera, m_HDShadowInitParameters, m_NewShadowManager, light, cullResults, lightIndex, out shadowRequestCount);
+                                shadowIndex = additionalLightData.UpdateShadowRequest(camera, m_ShadowInitParameters, m_ShadowManager, light, cullResults, lightIndex, out shadowRequestCount);
 
 #if UNITY_EDITOR
                                 if (debugDisplaySettings.lightingDebugSettings.shadowDebugUseSelection && UnityEditor.Selection.activeGameObject == lightComponent.gameObject)
@@ -1992,7 +1710,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
                         // Directional rendering side, it is separated as it is always visible so no volume to handle here
                         if (gpuLightType == GPULightType.Directional)
                         {
-                            if (GetDirectionalLightData(cmd, shadowSettings, gpuLightType, light, lightComponent, additionalLightData, additionalShadowData, lightIndex, shadowIndex, debugDisplaySettings, directionalLightcount))
+                            if (GetDirectionalLightData(cmd, gpuLightType, light, lightComponent, additionalLightData, additionalShadowData, lightIndex, shadowIndex, debugDisplaySettings, directionalLightcount))
                             {
                                 directionalLightcount++;
 
@@ -2013,7 +1731,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
                         Vector3 lightDimensions = new Vector3(); // X = length or width, Y = height, Z = range (depth)
 
                         // Punctual, area, projector lights - the rendering side.
-                        if (GetLightData(cmd, shadowSettings, camera, gpuLightType, light, lightComponent, additionalLightData, additionalShadowData, lightIndex, shadowIndex, ref lightDimensions, debugDisplaySettings))
+                        if (GetLightData(cmd, hdShadowSettings, camera, gpuLightType, light, lightComponent, additionalLightData, additionalShadowData, lightIndex, shadowIndex, ref lightDimensions, debugDisplaySettings))
                         {
                             switch (lightCategory)
                             {
@@ -2046,10 +1764,8 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
                         }
                     }
                     
-                    if (useNewShadowSystem)
-                    {
-                        m_NewShadowManager.ProcessShadowRequests(cullResults, camera, debugDisplaySettings.lightingDebugSettings);
-                    }
+                    // Prepare the shadow datas for GPU and layout the shadow atlases
+                    m_ShadowManager.ProcessShadowRequests(cullResults, camera, debugDisplaySettings.lightingDebugSettings);
 
                     //Activate contact shadows on dominant light
                     if (m_DominantLightIndex != -1)
@@ -2571,10 +2287,8 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
                 Camera camera = hdCamera.camera;
 
                 // Shadows
-                m_NewShadowManager.SyncData();
-                m_NewShadowManager.BindResources(cmd);
-                m_ShadowMgr.SyncData();
-                m_ShadowMgr.BindResources(cmd, null, 0);
+                m_ShadowManager.SyncData();
+                m_ShadowManager.BindResources(cmd);
 
                 cmd.SetGlobalTexture(HDShaderIDs._CookieTextures, m_CookieTexArray.GetTexCache());
                 cmd.SetGlobalTexture(HDShaderIDs._CookieCubeTextures, m_CubeCookieTexArray.GetTexCache());
@@ -2626,10 +2340,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
         public void RenderShadows(ScriptableRenderContext renderContext, CommandBuffer cmd, CullResults cullResults)
         {
             // kick off the shadow jobs here
-            if (useNewShadowSystem)
-                m_NewShadowManager.RenderShadows(renderContext, cmd, cullResults);
-            else
-                m_ShadowMgr.RenderShadows(m_FrameId, renderContext, cmd, cullResults, cullResults.visibleLights);
+            m_ShadowManager.RenderShadows(renderContext, cmd, cullResults);
         }
 
         public struct LightingPassOptions
@@ -2679,9 +2390,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
                     lightDirection.w = 0;
                 }
 
-                m_NewShadowManager.BindResources(cmd);
-                // Passing the compute shader and it's kernel is totally useless (because the binder in ShadowSetup ignore them)
-                m_ShadowMgr.BindResources(cmd, screenSpaceShadowComputeShader, kernel);
+                m_ShadowManager.BindResources(cmd);
 
                 float contactShadowRange = Mathf.Clamp(m_ContactShadows.fadeDistance, 0.0f, m_ContactShadows.maxDistance);
                 float contactShadowFadeEnd = m_ContactShadows.maxDistance;
@@ -2928,72 +2637,31 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
 
             using (new ProfilingSample(cmd, "Display Shadows", CustomSamplerId.TPDisplayShadows.GetSampler()))
             {
-                if (!useNewShadowSystem)
+                if (lightingDebug.shadowDebugMode == ShadowMapDebugMode.VisualizeShadowMap)
                 {
-                    if (lightingDebug.shadowDebugMode == ShadowMapDebugMode.VisualizeShadowMap)
+                    int startShadowIndex = (int)lightingDebug.shadowMapIndex;
+                    int shadowRequestCount = 1;
+
+#if UNITY_EDITOR
+                    if (lightingDebug.shadowDebugUseSelection)
                     {
-                        int index = (int)lightingDebug.shadowMapIndex;
-
-    #if UNITY_EDITOR
-                        if (lightingDebug.shadowDebugUseSelection)
-                        {
-                            index = -1;
-                            if (UnityEditor.Selection.activeObject is GameObject)
-                            {
-                                GameObject go = UnityEditor.Selection.activeObject as GameObject;
-                                Light light = go.GetComponent<Light>();
-                                if (light != null)
-                                {
-                                    index = m_ShadowMgr.GetShadowRequestIndex(light);
-                                }
-                            }
-                        }
-    #endif
-
-                        if (index != -1)
-                        {
-                            uint faceCount = m_ShadowMgr.GetShadowRequestFaceCount((uint)index);
-                            for (uint i = 0; i < faceCount; ++i)
-                            {
-                                m_ShadowMgr.DisplayShadow(cmd, m_DebugShadowMapMaterial, index, i, x, y, overlaySize, overlaySize, lightingDebug.shadowMinValue, lightingDebug.shadowMaxValue, hdCamera.camera.cameraType != CameraType.SceneView);
-                                HDUtils.NextOverlayCoord(ref x, ref y, overlaySize, overlaySize, hdCamera.actualWidth);
-                            }
-                        }
+                        startShadowIndex = m_DebugSelectedLightShadowIndex;
+                        shadowRequestCount = m_DebugSelectedLightShadowCount;
                     }
-                    else if (lightingDebug.shadowDebugMode == ShadowMapDebugMode.VisualizeAtlas)
+#endif
+
+                    for (int shadowIndex = startShadowIndex; shadowIndex < startShadowIndex + shadowRequestCount; shadowIndex++)
                     {
-                        m_ShadowMgr.DisplayShadowMap(cmd, m_DebugShadowMapMaterial, lightingDebug.shadowAtlasIndex, lightingDebug.shadowSliceIndex, x, y, overlaySize, overlaySize, lightingDebug.shadowMinValue, lightingDebug.shadowMaxValue, hdCamera.camera.cameraType != CameraType.SceneView);
+                        m_ShadowManager.DisplayShadowMap(shadowIndex, cmd, m_DebugHDShadowMapMaterial, x, y, overlaySize, overlaySize, lightingDebug.shadowMinValue, lightingDebug.shadowMaxValue, hdCamera.camera.cameraType != CameraType.SceneView);
                         HDUtils.NextOverlayCoord(ref x, ref y, overlaySize, overlaySize, hdCamera.actualWidth);
                     }
                 }
-                else
+                else if (lightingDebug.shadowDebugMode == ShadowMapDebugMode.VisualizeAtlas)
                 {
-                    if (lightingDebug.shadowDebugMode == ShadowMapDebugMode.VisualizeShadowMap)
-                    {
-                        int startShadowIndex = (int)lightingDebug.shadowMapIndex;
-                        int shadowRequestCount = 1;
-
-#if UNITY_EDITOR
-                        if (lightingDebug.shadowDebugUseSelection)
-                        {
-                            startShadowIndex = m_DebugSelectedLightShadowIndex;
-                            shadowRequestCount = m_DebugSelectedLightShadowCount;
-                        }
-#endif
-
-                        for (int shadowIndex = startShadowIndex; shadowIndex < startShadowIndex + shadowRequestCount; shadowIndex++)
-                        {
-                            m_NewShadowManager.DisplayShadowMap(shadowIndex, cmd, m_DebugHDShadowMapMaterial, x, y, overlaySize, overlaySize, lightingDebug.shadowMinValue, lightingDebug.shadowMaxValue, hdCamera.camera.cameraType != CameraType.SceneView);
-                            HDUtils.NextOverlayCoord(ref x, ref y, overlaySize, overlaySize, hdCamera.actualWidth);
-                        }
-                    }
-                    else if (lightingDebug.shadowDebugMode == ShadowMapDebugMode.VisualizeAtlas)
-                    {
-                        m_NewShadowManager.DisplayShadowAtlas(cmd, m_DebugHDShadowMapMaterial, x, y, overlaySize, overlaySize, lightingDebug.shadowMinValue, lightingDebug.shadowMaxValue, hdCamera.camera.cameraType != CameraType.SceneView);
-                        HDUtils.NextOverlayCoord(ref x, ref y, overlaySize, overlaySize, hdCamera.actualWidth);
-                        m_NewShadowManager.DisplayShadowCascadeAtlas(cmd, m_DebugHDShadowMapMaterial, x, y, overlaySize, overlaySize, lightingDebug.shadowMinValue, lightingDebug.shadowMaxValue, hdCamera.camera.cameraType != CameraType.SceneView);
-                        HDUtils.NextOverlayCoord(ref x, ref y, overlaySize, overlaySize, hdCamera.actualWidth);
-                    }
+                    m_ShadowManager.DisplayShadowAtlas(cmd, m_DebugHDShadowMapMaterial, x, y, overlaySize, overlaySize, lightingDebug.shadowMinValue, lightingDebug.shadowMaxValue, hdCamera.camera.cameraType != CameraType.SceneView);
+                    HDUtils.NextOverlayCoord(ref x, ref y, overlaySize, overlaySize, hdCamera.actualWidth);
+                    m_ShadowManager.DisplayShadowCascadeAtlas(cmd, m_DebugHDShadowMapMaterial, x, y, overlaySize, overlaySize, lightingDebug.shadowMinValue, lightingDebug.shadowMaxValue, hdCamera.camera.cameraType != CameraType.SceneView);
+                    HDUtils.NextOverlayCoord(ref x, ref y, overlaySize, overlaySize, hdCamera.actualWidth);
                 }
             }
