+++ com.unity.render-pipelines.high-definition/Runtime/Lighting/LightLoop/LightLoopDef.hlsl
@@ -8,12 +8,7 @@ struct LightLoopContext
 {
     int sampleReflection;
 
-// TODO: remove once hd shadow system is stable
-#ifndef USE_CORE_SHADOW_SYSTEM
     HDShadowContext shadowContext;
-#else
-    ShadowContext shadowContext;
-#endif
     
     float contactShadow; // Currently we support only one contact shadow per view
     float shadowValue; // Stores the value of the cascade shadow map
