--- com.unity.render-pipelines.high-definition/Runtime/Lighting/LightLoop/Shadow.hlsl.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: a17db9285d434b8aafe02c6d8af3353c
-ShaderImporter:
-  externalObjects: {}
-  defaultTextures: []
-  nonModifiableTextures: []
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
