--- com.unity.render-pipelines.high-definition/Runtime/Lighting/Shadow/DebugDisplayShadowMap.shader.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: ee25e539f5594f44085e0a9000c15d9b
-timeCreated: 1476053153
-licenseType: Pro
-ShaderImporter:
-  defaultTextures: []
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
