+++ com.unity.render-pipelines.high-definition/Runtime/Lighting/Shadow/HDShadowUtils.cs
@@ -8,6 +8,11 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
     // TODO remove every occurrence of ShadowSplitData in function parameters when we'll have scriptable culling
     public static class HDShadowUtils
     {
+        public static float Asfloat(uint val) { unsafe { return *((float*)&val); } }
+        public static float Asfloat(int val)  { unsafe { return *((float*)&val); } }
+        public static int Asint(float val)    { unsafe { return *((int*)&val); } }
+        public static uint Asuint(float val)  { unsafe { return *((uint*)&val); } }
+
         static float GetPunctualFilterWidthInTexels(LightType lightType)
         {
             var hdAsset = (GraphicsSettings.renderPipelineAsset as HDRenderPipelineAsset);
@@ -31,8 +36,8 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
         {
             Vector4 lightDir;
 
-            float guardAngle = ShadowUtils.CalcGuardAnglePerspective(90.0f, viewportSize.x, GetPunctualFilterWidthInTexels(lightType), normalBiasMax, 79.0f);
-            ShadowUtils.ExtractPointLightMatrix(visibleLight, faceIndex, guardAngle, out view, out projection, out deviceProjection, out invViewProjection, out lightDir, out splitData);
+            float guardAngle = CalcGuardAnglePerspective(90.0f, viewportSize.x, GetPunctualFilterWidthInTexels(lightType), normalBiasMax, 79.0f);
+            ExtractPointLightMatrix(visibleLight, faceIndex, guardAngle, out view, out projection, out deviceProjection, out invViewProjection, out lightDir, out splitData);
         }
 
         // TODO: box spot and pyramid spots with non 1 aspect ratios shadow are incorrectly culled, see when scriptable culling will be here
@@ -44,8 +49,8 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
             if (shape != SpotLightShape.Pyramid)
                 aspectRatio = 1.0f;
 
-            float guardAngle = ShadowUtils.CalcGuardAnglePerspective(visibleLight.light.spotAngle, viewportSize.x, GetPunctualFilterWidthInTexels(lightType), normalBiasMax, 180.0f - visibleLight.light.spotAngle);
-            ShadowUtils.ExtractSpotLightMatrix(visibleLight, guardAngle, aspectRatio, out view, out projection, out deviceProjection, out invViewProjection, out lightDir, out splitData);
+            float guardAngle = CalcGuardAnglePerspective(visibleLight.light.spotAngle, viewportSize.x, GetPunctualFilterWidthInTexels(lightType), normalBiasMax, 180.0f - visibleLight.light.spotAngle);
+            ExtractSpotLightMatrix(visibleLight, guardAngle, aspectRatio, out view, out projection, out deviceProjection, out invViewProjection, out lightDir, out splitData);
 
             if (shape == SpotLightShape.Box)
             {
@@ -53,7 +58,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
                 float nearZ = visibleLight.light.shadowNearPlane >= nearMin ? visibleLight.light.shadowNearPlane : nearMin;
                 projection = Matrix4x4.Ortho(-shapeWidth / 2, shapeWidth / 2, -shapeHeight / 2, shapeHeight / 2, nearZ, visibleLight.range);
                 deviceProjection = GL.GetGPUProjectionMatrix(projection, false);
-                ShadowUtils.InvertOrthographic(ref projection, ref view, out invViewProjection);
+                InvertOrthographic(ref projection, ref view, out invViewProjection);
             }
         }
 
@@ -61,7 +66,22 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
         {
             Vector4     lightDir;
 
-            ShadowUtils.ExtractDirectionalLightMatrix(visibleLight, cascadeIndex, cascadeCount, cascadeRatios, nearPlaneOffset, (uint)viewportSize.x, (uint)viewportSize.y, out view, out projection, out deviceProjection, out invViewProjection, out lightDir, out splitData, cullResults, lightIndex);
+            Debug.Assert((uint)viewportSize.x == (uint)viewportSize.y, "Currently the cascaded shadow mapping code requires square cascades.");
+            splitData = new ShadowSplitData();
+            splitData.cullingSphere.Set(0.0f, 0.0f, 0.0f, float.NegativeInfinity);
+            splitData.cullingPlaneCount = 0;
+            // get lightDir
+            lightDir = visibleLight.light.transform.forward;
+            // TODO: At some point this logic should be moved to C#, then the parameters cullResults and lightIndex can be removed as well
+            //       For directional lights shadow data is extracted from the cullResults, so that needs to be somehow provided here.
+            //       Check ScriptableShadowsUtility.cpp ComputeDirectionalShadowMatricesAndCullingPrimitives(...) for details.
+            Vector3 ratios = new Vector3();
+            for (int i = 0, cnt = cascadeRatios.Length < 3 ? cascadeRatios.Length : 3; i < cnt; i++)
+                ratios[i] = cascadeRatios[i];
+            cullResults.ComputeDirectionalShadowMatricesAndCullingPrimitives(lightIndex, (int)cascadeIndex, cascadeCount, ratios, (int)viewportSize.x, nearPlaneOffset, out view, out projection, out splitData);
+            // and the compound (deviceProjection will potentially inverse-Z)
+            deviceProjection = GL.GetGPUProjectionMatrix(projection, false);
+            InvertOrthographic(ref deviceProjection, ref view, out invViewProjection);
         }
 
         // Currently area light shadows are not supported
@@ -72,6 +92,217 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
             deviceProjection = Matrix4x4.identity;
             projection = Matrix4x4.identity;
             splitData = default(ShadowSplitData);
+        }
+
+        public enum CubemapEdge
+        {
+            kCubeEdgePX_PY = 0,
+            kCubeEdgePX_NY,
+            kCubeEdgePX_PZ,
+            kCubeEdgePX_NZ,
+
+            kCubeEdgeNX_PY,
+            kCubeEdgeNX_NY,
+            kCubeEdgeNX_PZ,
+            kCubeEdgeNX_NZ,
+
+            kCubeEdgePY_PZ,
+            kCubeEdgePY_NZ,
+            kCubeEdgeNY_PZ,
+            kCubeEdgeNY_NZ,
+            kCubeEdge_Count
+        };
+
+        public static readonly CubemapEdge[,] kCubemapEdgesPerFace = new CubemapEdge[6, 4]
+        {
+            { CubemapEdge.kCubeEdgePX_PY, CubemapEdge.kCubeEdgePX_NY, CubemapEdge.kCubeEdgePX_PZ, CubemapEdge.kCubeEdgePX_NZ }, // PX
+            { CubemapEdge.kCubeEdgeNX_PY, CubemapEdge.kCubeEdgeNX_NY, CubemapEdge.kCubeEdgeNX_PZ, CubemapEdge.kCubeEdgeNX_NZ }, // NX
+            { CubemapEdge.kCubeEdgePX_PY, CubemapEdge.kCubeEdgeNX_PY, CubemapEdge.kCubeEdgePY_PZ, CubemapEdge.kCubeEdgePY_NZ }, // PY
+            { CubemapEdge.kCubeEdgePX_NY, CubemapEdge.kCubeEdgeNX_NY, CubemapEdge.kCubeEdgeNY_PZ, CubemapEdge.kCubeEdgeNY_NZ }, // NY
+            { CubemapEdge.kCubeEdgePX_PZ, CubemapEdge.kCubeEdgeNX_PZ, CubemapEdge.kCubeEdgePY_PZ, CubemapEdge.kCubeEdgeNY_PZ }, // PZ
+            { CubemapEdge.kCubeEdgePX_NZ, CubemapEdge.kCubeEdgeNX_NZ, CubemapEdge.kCubeEdgePY_NZ, CubemapEdge.kCubeEdgeNY_NZ }  // NZ
+        };
+
+        const float oneOverSqr2 = 0.70710678118654752440084436210485f;
+        public static readonly Vector3[] kCubemapEdgeDirections = new Vector3[(int)CubemapEdge.kCubeEdge_Count]
+        {
+            new Vector3(oneOverSqr2,  oneOverSqr2,            0),
+            new Vector3(oneOverSqr2, -oneOverSqr2,            0),
+            new Vector3(oneOverSqr2,            0,  oneOverSqr2),
+            new Vector3(oneOverSqr2,            0, -oneOverSqr2),
+
+            new Vector3(-oneOverSqr2,  oneOverSqr2,            0),
+            new Vector3(-oneOverSqr2, -oneOverSqr2,            0),
+            new Vector3(-oneOverSqr2,            0,  oneOverSqr2),
+            new Vector3(-oneOverSqr2,            0, -oneOverSqr2),
+
+            new Vector3(0,  oneOverSqr2,  oneOverSqr2),
+            new Vector3(0,  oneOverSqr2, -oneOverSqr2),
+            new Vector3(0, -oneOverSqr2,  oneOverSqr2),
+            new Vector3(0, -oneOverSqr2, -oneOverSqr2)
+        };
+
+        // Cubemap faces with flipped z coordinate.
+        // These matrices do NOT match what we have in Skybox.cpp.
+        // The C++ runtime flips y as well and requires patching up
+        // the culling state. Using these matrices keeps the winding
+        // order, but may need some special treatment if rendering
+        // into an actual cubemap.
+        public static readonly Matrix4x4[] kCubemapFaces = new Matrix4x4[]
+        {
+            new Matrix4x4( // pos X
+                new Vector4(0.0f,  0.0f, -1.0f,  0.0f),
+                new Vector4(0.0f,  1.0f,  0.0f,  0.0f),
+                new Vector4(-1.0f,  0.0f,  0.0f,  0.0f),
+                new Vector4(0.0f,  0.0f,  0.0f,  1.0f)),
+            new Matrix4x4( // neg x
+                new Vector4(0.0f,  0.0f,  1.0f,  0.0f),
+                new Vector4(0.0f,  1.0f,  0.0f,  0.0f),
+                new Vector4(1.0f,  0.0f,  0.0f,  0.0f),
+                new Vector4(0.0f,  0.0f,  0.0f,  1.0f)),
+            new Matrix4x4( // pos y
+                new Vector4(1.0f,  0.0f,  0.0f,  0.0f),
+                new Vector4(0.0f,  0.0f, -1.0f,  0.0f),
+                new Vector4(0.0f, -1.0f,  0.0f,  0.0f),
+                new Vector4(0.0f,  0.0f,  0.0f,  1.0f)),
+            new Matrix4x4( // neg y
+                new Vector4(1.0f,  0.0f,  0.0f,  0.0f),
+                new Vector4(0.0f,  0.0f,  1.0f,  0.0f),
+                new Vector4(0.0f,  1.0f,  0.0f,  0.0f),
+                new Vector4(0.0f,  0.0f,  0.0f,  1.0f)),
+            new Matrix4x4( // pos z
+                new Vector4(1.0f,  0.0f,  0.0f,  0.0f),
+                new Vector4(0.0f,  1.0f,  0.0f,  0.0f),
+                new Vector4(0.0f,  0.0f, -1.0f,  0.0f),
+                new Vector4(0.0f,  0.0f,  0.0f,  1.0f)),
+            new Matrix4x4( // neg z
+                new Vector4(-1.0f,  0.0f,  0.0f,  0.0f),
+                new Vector4(0.0f,  1.0f,  0.0f,  0.0f),
+                new Vector4(0.0f,  0.0f,  1.0f,  0.0f),
+                new Vector4(0.0f,  0.0f,  0.0f,  1.0f))
+        };
+
+        static void InvertView(ref Matrix4x4 view, out Matrix4x4 invview)
+        {
+            invview = Matrix4x4.zero;
+            invview.m00 = view.m00; invview.m01 = view.m10; invview.m02 = view.m20;
+            invview.m10 = view.m01; invview.m11 = view.m11; invview.m12 = view.m21;
+            invview.m20 = view.m02; invview.m21 = view.m12; invview.m22 = view.m22;
+            invview.m33 = 1.0f;
+            invview.m03 = -(invview.m00 * view.m03 + invview.m01 * view.m13 + invview.m02 * view.m23);
+            invview.m13 = -(invview.m10 * view.m03 + invview.m11 * view.m13 + invview.m12 * view.m23);
+            invview.m23 = -(invview.m20 * view.m03 + invview.m21 * view.m13 + invview.m22 * view.m23);
+        }
+
+        static void InvertOrthographic(ref Matrix4x4 proj, ref Matrix4x4 view, out Matrix4x4 vpinv)
+        {
+            Matrix4x4 invview;
+            InvertView(ref view, out invview);
+
+            Matrix4x4 invproj = Matrix4x4.zero;
+            invproj.m00 = 1.0f / proj.m00;
+            invproj.m11 = 1.0f / proj.m11;
+            invproj.m22 = 1.0f / proj.m22;
+            invproj.m33 = 1.0f;
+            invproj.m03 =   proj.m03 * invproj.m00;
+            invproj.m13 =   proj.m13 * invproj.m11;
+            invproj.m23 = -proj.m23 * invproj.m22;
+
+            vpinv = invview * invproj;
+        }
+
+        static void InvertPerspective(ref Matrix4x4 proj, ref Matrix4x4 view, out Matrix4x4 vpinv)
+        {
+            Matrix4x4 invview;
+            InvertView(ref view, out invview);
+
+            Matrix4x4 invproj = Matrix4x4.zero;
+            invproj.m00 = 1.0f / proj.m00;
+            invproj.m03 = proj.m02 * invproj.m00;
+            invproj.m11 = 1.0f / proj.m11;
+            invproj.m13 = proj.m12 * invproj.m11;
+            invproj.m22 = 0.0f;
+            invproj.m23 = -1.0f;
+            invproj.m33 = proj.m22 / proj.m23;
+            invproj.m32 = invproj.m33 / proj.m22;
+
+            vpinv = invview * invproj;
+        }
+
+        static Matrix4x4 ExtractSpotLightMatrix(VisibleLight vl, float guardAngle, float aspectRatio, out Matrix4x4 view, out Matrix4x4 proj, out Matrix4x4 deviceProj, out Matrix4x4 vpinverse, out Vector4 lightDir, out ShadowSplitData splitData)
+        {
+            splitData = new ShadowSplitData();
+            splitData.cullingSphere.Set(0.0f, 0.0f, 0.0f, float.NegativeInfinity);
+            splitData.cullingPlaneCount = 0;
+            // get lightDir
+            lightDir = vl.light.transform.forward;
+            // calculate view
+            Matrix4x4 scaleMatrix = Matrix4x4.identity;
+            scaleMatrix.m22 = -1.0f;
+            view = scaleMatrix * vl.localToWorld.inverse;
+            // following code is from SharedLightData::GetNearPlaneMinBound
+            float percentageBound = 0.01f * vl.light.range;
+            float fixedBound = 0.1f;
+            float nearmin = fixedBound <= percentageBound ? fixedBound : percentageBound;
+            // calculate projection
+            float zfar = vl.range;
+            float znear = vl.light.shadowNearPlane >= nearmin ? vl.light.shadowNearPlane : nearmin;
+            float fov = vl.spotAngle + guardAngle;
+            proj = Matrix4x4.Perspective(fov, aspectRatio, znear, zfar);
+            // and the compound (deviceProj will potentially inverse-Z)
+            deviceProj = GL.GetGPUProjectionMatrix(proj, false);
+            InvertPerspective(ref deviceProj, ref view, out vpinverse);
+            return deviceProj * view;
+        }
+
+        static Matrix4x4 ExtractPointLightMatrix(VisibleLight vl, uint faceIdx, float guardAngle, out Matrix4x4 view, out Matrix4x4 proj, out Matrix4x4 deviceProj, out Matrix4x4 vpinverse, out Vector4 lightDir, out ShadowSplitData splitData)
+        {
+            if (faceIdx > (uint)CubemapFace.NegativeZ)
+                Debug.LogError("Tried to extract cubemap face " + faceIdx + ".");
+
+            splitData = new ShadowSplitData();
+            splitData.cullingSphere.Set(0.0f, 0.0f, 0.0f, float.NegativeInfinity);
+            splitData.cullingPlaneCount = 4;
+            // get lightDir
+            lightDir = vl.light.transform.forward;
+            // calculate the view matrices
+            Vector3 lpos = vl.light.transform.position;
+            view = kCubemapFaces[faceIdx];
+            Vector3 inverted_viewpos = kCubemapFaces[faceIdx].MultiplyPoint(-lpos);
+            view.SetColumn(3, new Vector4(inverted_viewpos.x, inverted_viewpos.y, inverted_viewpos.z, 1.0f));
+
+            for (int i = 0; i < 4; ++i)
+            {
+                CubemapEdge cubemapEdge = kCubemapEdgesPerFace[faceIdx, i];
+                Vector3 cullingPlaneDirection = kCubemapEdgeDirections[(int)cubemapEdge];
+                splitData.SetCullingPlane(i, new Plane(cullingPlaneDirection, lpos));
+            }
+            // following code is from SharedLightData::GetNearPlaneMinBound
+            float percentageBound = 0.01f * vl.light.range;
+            float fixedBound = 0.1f;
+            float nearmin = fixedBound <= percentageBound ? fixedBound : percentageBound;
+            // calculate projection
+            float farPlane = vl.range;
+            float nearPlane = vl.light.shadowNearPlane >= nearmin ? vl.light.shadowNearPlane : nearmin;
+            proj = Matrix4x4.Perspective(90.0f + guardAngle, 1.0f, nearPlane, farPlane);
+            // and the compound (deviceProj will potentially inverse-Z)
+            deviceProj = GL.GetGPUProjectionMatrix(proj, false);
+            InvertPerspective(ref deviceProj, ref view, out vpinverse);
+            return deviceProj * view;
+        }
+
+        static float CalcGuardAnglePerspective(float angleInDeg, float resolution, float filterWidth, float normalBiasMax, float guardAngleMaxInDeg)
+        {
+            float angleInRad  = angleInDeg * 0.5f * Mathf.Deg2Rad;
+            float res         = 2.0f / resolution;
+            float texelSize   = Mathf.Cos(angleInRad) * res;
+            float beta        = normalBiasMax * texelSize * 1.4142135623730950488016887242097f;
+            float guardAngle  = Mathf.Atan(beta);
+            texelSize   = Mathf.Tan(angleInRad + guardAngle) * res;
+            guardAngle  = Mathf.Atan((resolution + Mathf.Ceil(filterWidth)) * texelSize * 0.5f) * 2.0f * Mathf.Rad2Deg - angleInDeg;
+            guardAngle *= 2.0f;
+
+            return guardAngle < guardAngleMaxInDeg ? guardAngle : guardAngleMaxInDeg;
         }
     }
 }
