--- com.unity.render-pipelines.high-definition/Runtime/Lighting/Shadow/Shadow.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 11e59a3cb6482b245b20ac498cb3afd5
-timeCreated: 1491321441
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
