--- com.unity.render-pipelines.high-definition/Runtime/Lighting/Shadow/Shadow.hlsl.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 3d03971a8848ee14bb29a3c4ee9790b8
-timeCreated: 1491321444
-licenseType: Pro
-ShaderImporter:
-  defaultTextures: []
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
