--- com.unity.render-pipelines.high-definition/Runtime/Lighting/Shadow/ShadowAlgorithms.hlsl.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: e2ca1e297ddf0b54f9d72244a419844b
-timeCreated: 1491321445
-licenseType: Pro
-ShaderImporter:
-  defaultTextures: []
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
