--- com.unity.render-pipelines.high-definition/Runtime/Lighting/Shadow/ShadowBase.cs.hlsl.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 78929777594e2d84aba05af5bf6463d0
-timeCreated: 1491321444
-licenseType: Pro
-ShaderImporter:
-  defaultTextures: []
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
