--- com.unity.render-pipelines.high-definition/Runtime/Lighting/Shadow/ShadowBase.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 7640004f9f444ad40aba7003ccb31e7f
-timeCreated: 1491321441
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
