--- com.unity.render-pipelines.high-definition/Runtime/Lighting/Shadow/ShadowSampling.hlsl.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 870dbce0ad39569448fa01659ed39fd0
-ShaderImporter:
-  externalObjects: {}
-  defaultTextures: []
-  nonModifiableTextures: []
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
