--- com.unity.render-pipelines.high-definition/Runtime/Lighting/Shadow/ShadowTexFetch.hlsl.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: a9c17fbca36dfb744b42fb48d3829887
-timeCreated: 1491321444
-licenseType: Pro
-ShaderImporter:
-  defaultTextures: []
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
