--- com.unity.render-pipelines.high-definition/Runtime/Lighting/Shadow/ShadowUtilities.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: d53d8bd422ecef740865c3acd7e2c62b
-timeCreated: 1491321441
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
