--- com.unity.render-pipelines.high-definition/Runtime/Lighting/Shadow/VectorArray.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: b6860f4fd5236f648b2c48c6799c68a7
-timeCreated: 1491321441
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
