+++ com.unity.render-pipelines.high-definition/Runtime/Material/LayeredLit/LayeredLit.shader
@@ -719,10 +719,6 @@ Shader "HDRenderPipeline/LayeredLit"
             // Setup DECALS_OFF so the shader stripper can remove variants
             #pragma multi_compile DECALS_OFF DECALS_3RT DECALS_4RT
             
-            // TODO: remove this once new shadow system works
-            // Only for dev/test purpose, allow to switch dynamically between HD and Core shadow system
-            // #pragma multi_compile _ USE_CORE_SHADOW_SYSTEM
-            
             // Supported shadow modes per light type
             #pragma multi_compile PUNCTUAL_SHADOW_LOW PUNCTUAL_SHADOW_MEDIUM PUNCTUAL_SHADOW_HIGH
             #pragma multi_compile DIRECTIONAL_SHADOW_LOW DIRECTIONAL_SHADOW_MEDIUM DIRECTIONAL_SHADOW_HIGH
