+++ com.unity.render-pipelines.high-definition/Runtime/Material/StackLit/StackLit.shader
@@ -574,10 +574,6 @@ Shader "HDRenderPipeline/StackLit"
             #pragma multi_compile _ SHADOWS_SHADOWMASK
             // Setup DECALS_OFF so the shader stripper can remove variants
             #pragma multi_compile DECALS_OFF DECALS_3RT DECALS_4RT
-
-            // TODO: remove this once new shadow system works
-            // Only for dev/test purpose, allow to switch dynamically between HD and Core shadow system
-            // #pragma multi_compile _ USE_CORE_SHADOW_SYSTEM
             
             // Supported shadow modes per light type
             #pragma multi_compile PUNCTUAL_SHADOW_LOW PUNCTUAL_SHADOW_MEDIUM PUNCTUAL_SHADOW_HIGH
